One large overdue commit.
Added MobileClient to hold the orchestrator's view of it's mobile application clients.
Added MobileClientRegistry to hold all MobileClients.
Added ServiceNode to hold the orchestrator's view of it's edge-cloud service nodes.
Added ServiceNodeRegistry to hold all ServiceNodes.
Created the Trigger interface.
Created the MinRequirementsTrigger, empty LatencyTrigger.
Created the Migrator interface.
Created the Selector interface.
Slight edit to one of the orchestrator.properties.
Removed rollingAverage as a parameter to the Orchestrator constructor.
Added the ServiceNodeRegistry and MobileClientRegistry to Orchestrator.
Added a warning comment to Orchestrator to illustrate the problem of receiving new clients (line 118).
Other minor changes to Orchestrator.
Minor changes to orchestrator.properties.
Added a todo to AbstractServiceNode.
Added a javadoc comment to explain the purpose of ServiceHost.
Added a todo to NodeInfo. NodeInfo has been replaced by the ServiceNode class for holding info. It should still be used as a Message object.

Signed-off-by: Joseph Dempsey <joseph.dempsey@ucdconnect.ie>

Former-commit-id: ed2217f38ad9323b7495a8f36b92159d31843c3c