Made fields in ServiceRequest & ServiceResponse clearer.
Fixed an issue in Orchestrator.handleServiceResponse where the WebSocket over which to send the ServiceResponse was being searched for in the MobileClientRegistry.
The ServiceResponse needs to be sent to the migration Target, so, the WebSocket is in the ServiceNodeRegistry.
Added SimpleSelector for the LatencyTrigger.
Migrator: removed trigger and replaced with migrate.

Signed-off-by: Joseph Dempsey <joseph.dempsey@ucdconnect.ie>

Former-commit-id: de0305e2292d722afe13447e2be1ddf777b3b63d